# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjetBank API",
        "description": "API de gestion bancaire",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/v1/transactions/depot": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un dépôt",
                "description": "Process a deposit",
                "operationId": "b25357ddc41cafb1c1912f91ac368364",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id",
                                    "montant"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 500
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Dépôt de salaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dépôt effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "compte_id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "depot"
                                        },
                                        "montant": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 500
                                        },
                                        "devise": {
                                            "type": "string",
                                            "example": "FCFA"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Dépôt d'argent"
                                        },
                                        "statut": {
                                            "type": "string",
                                            "example": "validee"
                                        },
                                        "date_execution": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/v1/transactions/retrait": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un retrait",
                "description": "Process a withdrawal",
                "operationId": "fc085500db9f3f905f76c954bbd26e78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id",
                                    "montant"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 200
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Retrait DAB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retrait effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "compte_id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "retrait"
                                        },
                                        "montant": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 200
                                        },
                                        "devise": {
                                            "type": "string",
                                            "example": "FCFA"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Retrait d'argent"
                                        },
                                        "statut": {
                                            "type": "string",
                                            "example": "validee"
                                        },
                                        "date_execution": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solde insuffisant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/v1/transactions/virement": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un virement",
                "description": "Process a transfer",
                "operationId": "c59300a3fc8b17fcecdd262321190a2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_source_id",
                                    "compte_destination_id",
                                    "montant"
                                ],
                                "properties": {
                                    "compte_source_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid"
                                    },
                                    "compte_destination_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 300
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Virement vers compte épargne"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Virement effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "compte_id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "compte_destination_id": {
                                            "type": "string",
                                            "example": "uuid"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "virement"
                                        },
                                        "montant": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 300
                                        },
                                        "devise": {
                                            "type": "string",
                                            "example": "FCFA"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Virement bancaire"
                                        },
                                        "statut": {
                                            "type": "string",
                                            "example": "validee"
                                        },
                                        "date_execution": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solde insuffisant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}